+++ lib/core/film.h
@@ -57,8 +57,8 @@ public:
                                  int *ystart, int *yend) const = 0;
     virtual void GetPixelExtent(int *xstart, int *xend,
                                 int *ystart, int *yend) const = 0;
-    virtual void UpdateDisplay(int x0, int y0, int x1, int y1, float splatScale = 1.f);
-    virtual void WriteImage(float splatScale = 1.f) = 0;
+    virtual void UpdateDisplay(int x0, int y0, int x1, int y1, double splatScale = 1.f);
+    virtual void WriteImage(double splatScale = 1.f) = 0;
 
     // Film Public Data
     const int xResolution, yResolution;
